refactor(manager/gradle): group parser directives in separate files
